Removed temporary Server version
